fix: repair PR monitor adoption CI regressions
