+++ docs/REASON_CATALOG.md
@@ -333,6 +333,13 @@ This catalog documents common API/CLI/MCP failures, likely causes, and operator 
 **Related Command:** `awf service doctor`
 **Docs Link:** [https://docs.x.ai/build/cli](https://docs.x.ai/build/cli)
 
+### HOSTED_DELEGATION_NOT_CONFIGURED
+**Problem:** Hosted PR monitor adoption was requested, but hosted delegation settings are incomplete.
+**Likely Cause:** The AWF service is missing a hosted delegation base URL or bearer token configuration.
+**Operator Fix:** Configure the hosted delegation service environment, then retry the adoption request.
+**Related Command:** `awf workspace adopt-pr --execution hosted`
+**Docs Link:** [docs/REASON_CATALOG.md#hosted_delegation_not_configured](#hosted_delegation_not_configured)
+
 ### HOST_PORT_CONFLICT
 **Problem:** AWF rejected a workspace create or retry because a host port needed by the new workspace is already in use by another active or unreleased workspace.
 **Likely Cause:** Another workspace's profile services or companions bind the same Docker host port and its compose stack is still running, or the workspace is terminal but has not yet released its runtime resources (no `workspace.terminal_runtime_released` event exists). For auto-resolved profiles, the conflict may also be detected at provision time by the provisioner's host-port re-check (``_check_auto_resolved_profile_host_ports``) rather than at dispatch, surfacing as an ``INFRASTRUCTURE_FAILURE`` instead of a 409.
