+++ tests/unit/contracts/_capabilities.py
@@ -523,6 +523,7 @@ _CAPABILITIES: tuple[ContractCapability, ...] = (
                 "profile",
                 "owned_paths",
                 "auto_merge",
+                "execution",
                 "initial_review_grace_period_seconds",
                 "task_title",
                 "task_prompt",
@@ -543,6 +544,7 @@ _CAPABILITIES: tuple[ContractCapability, ...] = (
                 "profile",
                 "owned_paths",
                 "auto_merge",
+                "execution",
                 "initial_review_grace_period_seconds",
                 "task_title",
                 "task_prompt",
@@ -561,6 +563,7 @@ _CAPABILITIES: tuple[ContractCapability, ...] = (
                 "--owned-path",
                 "--profile",
                 "--auto-merge/--no-auto-merge",
+                "--execution",
                 "--initial-review-grace-period-seconds",
                 "--title",
                 "--prompt",
